Fix semantics of "Unvalidated" domains
This change allows multiple "policies" for a particular domain to exist
at once in the database. The reason we need this: If someone wants to
update or modify their policy while it's still in testing mode, the
actual update shouldn't occur until email confirmation has occurred for
the new policy. At the same time, we still need to remember this new
policy in case the update does occur.

People can still update their entry on the list so long as it's in "testing"
mode. Once it's in "enforce", it is much harder to change their entry on
the list, and for now, we have to go through a manual channel.

As a side thing, also removed "GET" functionality from /queue endpoint
since no one is consuming it or relying on it.